requirements.txt
fastapi==0.88.0
python-dotenv==0.21.0
uvicorn==0.20.0
psycopg2==2.9.5
sqlalchemy==1.4.41
email-validator==1.3.0
passlib[bcrypt]
python-jose[cryptography]
python-multipar==0.0.5